+++ inst/examples/gallery/app.R
@@ -17,7 +17,7 @@ sapply(
 # shiny app
 shinyApp(
   ui = f7Page(
-    allowPWA = TRUE, options = list(dark = TRUE, color = "red"),
+    allowPWA = TRUE, options = list(dark = TRUE, filled = FALSE),
     f7TabLayout(
       appbar = f7Appbar(
         f7Flex(f7Back(targetId = "tabset"), f7Next(targetId = "tabset")),
@@ -199,7 +199,12 @@ shinyApp(
     output$popupContent <- renderPrint(input$popupText)
 
     observeEvent(input$togglePopup, {
-      updateF7Popup(id = "popup1")
+      f7Popup(
+        id = "popup1",
+        title = "My first popup",
+        f7Text("popupText", "Popup content", "This is my first popup ever, I swear!"),
+        verbatimTextOutput("popupContent")
+      )
     })
 
     observeEvent(input$popup1, {